Finished day 19
Decided to try and follow Norvig's way. It's basically his code, but I
did make it somewhat different so that it wasn't a complete
plagiarism. I mainly went this route for two reasons. First, the regex
version didn't work well with the python regex package. Second, it
seemed like the regex solution was a hack and that doing it this way
would help me learn more about parsing.